+++ src/main/java/boblovespi/factoryautomation/common/item/tools/ToolMaterial.java
@@ -33,23 +33,38 @@ public class ToolMaterial implements Tier
 		 */
 
 	public static int WOOD = toolMiningLevelCat.wood;
-	public static Tier flintMaterial = TierSortingRegistry.registerTier(new ToolMaterial(WOOD, 20, 4.5f, 1.5f, 2, () -> Ingredient.of(Items.FLINT)),
-																		new ResourceLocation(MODID, "flint"), List.of(Tiers.WOOD), List.of(Tiers.STONE));
+
+	public static Tier badFlintMaterial = TierSortingRegistry.registerTier(
+			new ToolMaterial(WOOD, 20, 4.5f, 1.5f, 2, () -> Ingredient.of(Items.FLINT)),
+			new ResourceLocation(MODID, "flint"), List.of(Tiers.WOOD), List.of(Tiers.STONE));
+	public static Tier goodFlintMaterial = TierSortingRegistry.registerTier(
+			new ToolMaterial(WOOD, Tiers.WOOD.getUses(), Tiers.WOOD.getSpeed(), Tiers.WOOD.getAttackDamageBonus(),
+					Tiers.WOOD.getEnchantmentValue(), () -> Ingredient.of(Items.FLINT)),
+			new ResourceLocation(MODID, "good_flint"), List.of(badFlintMaterial), List.of(Tiers.STONE));
+
 	public static int STONE = toolMiningLevelCat.stone;
 	public static int COPPER = toolMiningLevelCat.copper;
 	public static int IRON = toolMiningLevelCat.iron;
 	public static int BRONZE = toolMiningLevelCat.bronze;
-	public static Tier bronzeMaterial = TierSortingRegistry.registerTier(new ToolMaterial(BRONZE, 600, 5.0F, 2.0F, 12, () -> Ingredient.of(FATags.ForgeItemTag("ingots/bronze"))),
-																				 new ResourceLocation(MODID, "bronze"), List.of(Tiers.IRON), List.of());
+
+	public static Tier bronzeMaterial = TierSortingRegistry.registerTier(
+			new ToolMaterial(BRONZE, 600, 5.0F, 2.0F, 12, () -> Ingredient.of(FATags.ForgeItemTag("ingots/bronze"))),
+			new ResourceLocation(MODID, "bronze"), List.of(Tiers.IRON), List.of());
+
 	public static int DIAMOND = toolMiningLevelCat.diamond;
 	public static int STEEL = toolMiningLevelCat.steel;
 	public static int NETHERITE = toolMiningLevelCat.netherite;
-	public static Tier steelMaterial = TierSortingRegistry.registerTier(new ToolMaterial(STEEL, 1920, 6.5f, 4, 3, () -> Ingredient.of(FATags.ForgeItemTag("ingots/steel"))),
-																				new ResourceLocation(MODID, "steel"), List.of(bronzeMaterial), List.of(Tiers.DIAMOND));
-	public static Tier copperMaterial = TierSortingRegistry.registerTier(new ToolMaterial(COPPER, 240, 3.5f, 1.5f, 5, () -> Ingredient.of(FATags.ForgeItemTag("ingots/copper"))),
-																				 new ResourceLocation(MODID, "copper"), List.of(Tiers.STONE), List.of(Tiers.IRON));
-	public static Tier voidsteelMaterial = TierSortingRegistry.registerTier(new ToolMaterial(NETHERITE, 2269, 20f, 6.5f, 25, () -> Ingredient.of(FATags.ForgeItemTag("ingots/voidsteel"))),
-																		 new ResourceLocation(MODID, "voidsteel"), List.of(Tiers.NETHERITE), List.of());
+
+	public static Tier steelMaterial = TierSortingRegistry.registerTier(
+			new ToolMaterial(STEEL, 1920, 6.5f, 4, 3, () -> Ingredient.of(FATags.ForgeItemTag("ingots/steel"))),
+			new ResourceLocation(MODID, "steel"), List.of(bronzeMaterial), List.of(Tiers.DIAMOND));
+	public static Tier copperMaterial = TierSortingRegistry.registerTier(
+			new ToolMaterial(COPPER, 240, 3.5f, 1.5f, 5, () -> Ingredient.of(FATags.ForgeItemTag("ingots/copper"))),
+			new ResourceLocation(MODID, "copper"), List.of(Tiers.STONE), List.of(Tiers.IRON));
+	public static Tier voidsteelMaterial = TierSortingRegistry.registerTier(
+			new ToolMaterial(NETHERITE, 2269, 20f, 6.5f, 25,
+					() -> Ingredient.of(FATags.ForgeItemTag("ingots/voidsteel"))),
+			new ResourceLocation(MODID, "voidsteel"), List.of(Tiers.NETHERITE), List.of());
 	//
 
 	private final int harvestLevel;